requirements.txt
flask==2.3.3
werkzeug==2.3.7
pypdf2==3.0.1
python-docx==0.8.11
sentence-transformers==2.2.2
scikit-learn==1.2.2
numpy==1.24.3
spacy==3.5.0
torch==2.0.1
transformers==4.30.2
huggingface-hub==0.16.4 